Update overview image
For any cloud based or bare metal based implementation of any simulation, real-time processing task e.g. ML training, monitoring etc. Simvue is generic enough to be able to handle the monitoring, alerting and provide a single dashboard to allow users keep track of the tasks in an efficient and automated manner